Ground truth pose screwed-up coord system?
I remove the hack that claims that it fix Stage's screwed-up coord system. But in fact is that hack that makes the ground truth wrong, so I need to amend in my app.
Maybe the hack used to fix something that has been lately fixed in the underlying stagelib. Or maybe I just misunderstood the problem.... feel free to just reject this pull request in that case.